refactor(geography_utils): prefix package and namespace with autoware
