+++ common/autoware_geography_utils/CMakeLists.txt
@@ -1,5 +1,5 @@
 cmake_minimum_required(VERSION 3.14)
-project(geography_utils)
+project(autoware_geography_utils)
 
 find_package(autoware_cmake REQUIRED)
 autoware_package()
@@ -12,13 +12,13 @@ find_path(GeographicLib_INCLUDE_DIR GeographicLib/Config.h
 set(GeographicLib_INCLUDE_DIRS ${GeographicLib_INCLUDE_DIR})
 find_library(GeographicLib_LIBRARIES NAMES Geographic)
 
-ament_auto_add_library(geography_utils SHARED
+ament_auto_add_library(${PROJECT_NAME} SHARED
   src/height.cpp
   src/projection.cpp
   src/lanelet2_projector.cpp
 )
 
-target_link_libraries(geography_utils
+target_link_libraries(${PROJECT_NAME}
   ${GeographicLib_LIBRARIES}
 )
 
@@ -27,10 +27,10 @@ if(BUILD_TESTING)
 
   file(GLOB_RECURSE test_files test/*.cpp)
 
-  ament_add_ros_isolated_gtest(test_geography_utils ${test_files})
+  ament_add_ros_isolated_gtest(test_${PROJECT_NAME} ${test_files})
 
-  target_link_libraries(test_geography_utils
-  geography_utils
+  target_link_libraries(test_${PROJECT_NAME}
+  ${PROJECT_NAME}
   )
 endif()
