+++ planning/autoware_static_centerline_generator/package.xml
@@ -16,6 +16,7 @@
   <build_depend>rosidl_default_generators</build_depend>
 
   <depend>autoware_behavior_path_planner_common</depend>
+  <depend>autoware_geography_utils</depend>
   <depend>autoware_lanelet2_extension</depend>
   <depend>autoware_map_msgs</depend>
   <depend>autoware_mission_planner</depend>
@@ -27,7 +28,6 @@
   <depend>autoware_route_handler</depend>
   <depend>autoware_universe_utils</depend>
   <depend>autoware_vehicle_info_utils</depend>
-  <depend>geography_utils</depend>
   <depend>geometry_msgs</depend>
   <depend>global_parameter_loader</depend>
   <depend>interpolation</depend>
