+++ sensing/gnss_poser/package.xml
@@ -19,12 +19,12 @@
 
   <build_depend>libboost-dev</build_depend>
 
+  <depend>autoware_geography_utils</depend>
   <depend>autoware_sensing_msgs</depend>
   <depend>component_interface_specs</depend>
   <depend>component_interface_utils</depend>
   <depend>geographic_msgs</depend>
   <depend>geographiclib</depend>
-  <depend>geography_utils</depend>
   <depend>geometry_msgs</depend>
   <depend>rclcpp</depend>
   <depend>rclcpp_components</depend>
