+++ system/default_ad_api/package.xml
@@ -15,6 +15,7 @@
   <depend>autoware_ad_api_specs</depend>
   <depend>autoware_adapi_v1_msgs</depend>
   <depend>autoware_adapi_version_msgs</depend>
+  <depend>autoware_geography_utils</depend>
   <depend>autoware_motion_utils</depend>
   <depend>autoware_planning_msgs</depend>
   <depend>autoware_system_msgs</depend>
@@ -24,7 +25,6 @@
   <depend>component_interface_utils</depend>
   <depend>diagnostic_graph_utils</depend>
   <depend>geographic_msgs</depend>
-  <depend>geography_utils</depend>
   <depend>nav_msgs</depend>
   <depend>rclcpp</depend>
   <depend>rclcpp_components</depend>
